Rename `teleport-plugin-ms-teams` to `teleport-plugin-msteams`
