+++ access/msteams/Makefile
@@ -2,25 +2,25 @@ VERSION=10.0.2
 GO_VERSION=1.18.6
 
 BUILDDIR ?= build
-BINARY = $(BUILDDIR)/teleport-ms-teams
+BINARY = $(BUILDDIR)/teleport-msteams
 GITTAG=v$(VERSION)
-GITREF ?= $(shell git describe --dirty --long --tags --match '*ms-teams*')
+GITREF ?= $(shell git describe --dirty --long --tags --match '*msteams*')
 ADDFLAGS ?=
 BUILDFLAGS ?= $(ADDFLAGS) -ldflags "-w -s -X main.Gitref=$(GITREF) -X main.Version=$(VERSION)"
 CGOFLAG ?= CGO_ENABLED=1
 
 OS ?= $(shell go env GOOS)
 ARCH ?= $(shell go env GOARCH)
-RELEASE_NAME=teleport-access-ms-teams
+RELEASE_NAME=teleport-access-msteams
 RELEASE=$(RELEASE_NAME)-$(GITTAG)-$(OS)-$(ARCH)-bin
 
 RELEASE_MESSAGE = "Building with GOOS=$(OS) GOARCH=$(ARCH)."
 
-DOCKER_NAME = teleport-plugin-ms-teams
+DOCKER_NAME = teleport-plugin-msteams
 DOCKER_IMAGE = 146628656107.dkr.ecr.us-west-2.amazonaws.com/gravitational/$(DOCKER_NAME):$(VERSION)
 DOCKER_IMAGE_ECR_PUBLIC = public.ecr.aws/gravitational/$(DOCKER_NAME):$(VERSION)
 DOCKER_IMAGE_QUAY = quay.io/gravitational/$(DOCKER_NAME):$(VERSION)
-DOCKER_BUILD_ARGS = --build-arg GO_VERSION=${GO_VERSION} --build-arg ACCESS_PLUGIN=ms-teams --build-arg GITREF=$(GITREF)
+DOCKER_BUILD_ARGS = --build-arg GO_VERSION=${GO_VERSION} --build-arg ACCESS_PLUGIN=msteams --build-arg GITREF=$(GITREF)
 
 .PHONY: $(BINARY)
 $(BINARY):
