+++ access/msteams/uninstall.go
