+++ charts/access/msteams/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v2
-name: teleport-plugin-ms-teams
+name: teleport-plugin-msteams
 description: A Helm chart for the Teleport MsTeams Plugin
 
 # A chart can be either an 'application' or a 'library' chart.
