+++ charts/access/msteams/README.md
@@ -7,10 +7,10 @@ This chart sets up and configures a Deployment for the Access Request MsTeams pl
 ### Prerequisites
 
 As the MsTeams setup requires to download the plugin locally to generate assets to load in MsTeams,
-you must follow [the MsTeams access request guide](https://goteleport.com/docs/access-controls/access-request-plugins/ssh-approval-ms-teams/).
+you must follow [the MsTeams access request guide](https://goteleport.com/docs/access-controls/access-request-plugins/ssh-approval-msteams/).
 When generating the identity file, choose the "proxy" solution that generates a single file `auth.pem`.
 
-Once the guide in finished, you should have a working `teleport-ms-teams.toml` configuration file.
+Once the guide in finished, you should have a working `teleport-msteams.toml` configuration file.
 
 Recover the following values from it:
 - msapi.appID
@@ -36,17 +36,17 @@ Run the following command to create the secret from the `auth.pem` file recovere
 
 ```shell
 export NAMESPACE="your-namespace" #The namespace should already exist
-kubectl create -n "$NAMESPACE" secret generic teleport-plugin-ms-teams-identity --from-file=auth_id=./auth.pem
+kubectl create -n "$NAMESPACE" secret generic teleport-plugin-msteams-identity --from-file=auth_id=./auth.pem
 ```
 
 ### Installing the chart
 
-Create the value file `teleport-plugin-ms-teams-values.yaml` with the following content:
+Create the value file `teleport-plugin-msteams-values.yaml` with the following content:
 
 ```yaml
 teleport:
   address: "YOUR-TELEPORT-ADDRESS"
-  identitySecretName: "teleport-plugin-ms-teams-identity"
+  identitySecretName: "teleport-plugin-msteams-identity"
 
 msTeams:
   appID: "YOUR-APPID"
@@ -65,5 +65,5 @@ The `roleToRecipient` map controls which channels and users will be notified if 
 Finally, create a release from the Helm chart with the values:
 
 ```shell
-helm install teleport-plugin-ms-teams teleport/teleport-plugin-ms-teams --values teleport-plugin-ms-teams-values.yaml -n "$NAMESPACE"
+helm install teleport-plugin-msteams teleport/teleport-plugin-msteams --values teleport-plugin-msteams-values.yaml -n "$NAMESPACE"
 ```
