+++ charts/access/msteams/templates/configmap.yaml
@@ -1,22 +1,22 @@
 apiVersion: v1
 kind: ConfigMap
 metadata:
-  name: {{ include "ms-teams.fullname" . }}
+  name: {{ include "msteams.fullname" . }}
   labels:
-    {{- include "ms-teams.labels" . | nindent 4 }}
+    {{- include "msteams.labels" . | nindent 4 }}
 data:
-  teleport-ms-teams.toml: |
+  teleport-msteams.toml: |
     preload = true
     
     [teleport]
     addr = "{{ .Values.teleport.address }}"
-    identity = "/var/lib/teleport/plugins/ms-teams/auth_id"
+    identity = "/var/lib/teleport/plugins/msteams/auth_id"
 
     [msapi]
     app_id = "{{ .Values.msTeams.appID }}"
     tenant_id = "{{ .Values.msTeams.tenantID }}"
     teams_app_id = "{{ .Values.msTeams.teamsAppID }}"
-    app_secret = "/var/lib/teleport/plugins/ms-teams/appSecret"
+    app_secret = "/var/lib/teleport/plugins/msteams/appSecret"
 
     [role_to_recipients]
     {{- range $role, $recipients := .Values.roleToRecipients }}
