+++ charts/access/msteams/templates/deployment.yaml
@@ -1,14 +1,14 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: {{ include "ms-teams.fullname" . }}
+  name: {{ include "msteams.fullname" . }}
   labels:
-    {{- include "ms-teams.labels" . | nindent 4 }}
+    {{- include "msteams.labels" . | nindent 4 }}
 spec:
   replicas: 1
   selector:
     matchLabels:
-      {{- include "ms-teams.selectorLabels" . | nindent 6 }}
+      {{- include "msteams.selectorLabels" . | nindent 6 }}
   template:
     metadata:
       {{- with .Values.podAnnotations }}
@@ -16,7 +16,7 @@ spec:
         {{- toYaml . | nindent 8 }}
       {{- end }}
       labels:
-        {{- include "ms-teams.selectorLabels" . | nindent 8 }}
+        {{- include "msteams.selectorLabels" . | nindent 8 }}
     spec:
       {{- with .Values.imagePullSecrets }}
       imagePullSecrets:
@@ -34,18 +34,18 @@ spec:
             - /usr/local/bin/teleport-plugin
             - start
             - "--config"
-            - "/etc/teleport-ms-teams.toml"
+            - "/etc/teleport-msteams.toml"
           resources:
             {{- toYaml .Values.resources | nindent 12 }}
           volumeMounts:
             - name: config
-              mountPath: /etc/teleport-ms-teams.toml
-              subPath: teleport-ms-teams.toml
+              mountPath: /etc/teleport-msteams.toml
+              subPath: teleport-msteams.toml
             - name: auth-id
-              mountPath: /var/lib/teleport/plugins/ms-teams/auth_id
+              mountPath: /var/lib/teleport/plugins/msteams/auth_id
               subPath: {{ .Values.teleport.identitySecretPath }}
             - name: {{ .Values.secretVolumeName }}
-              mountPath: /var/lib/teleport/plugins/ms-teams/appSecret
+              mountPath: /var/lib/teleport/plugins/msteams/appSecret
               subPath: {{ .Values.msTeams.appSecretFromSecretPath }}
       {{- with .Values.nodeSelector }}
       nodeSelector:
@@ -62,7 +62,7 @@ spec:
       volumes:
         - name: config
           configMap:
-            name: {{ include "ms-teams.fullname" . }}
+            name: {{ include "msteams.fullname" . }}
             defaultMode: 0600
         - name: auth-id
           secret:
@@ -70,5 +70,5 @@ spec:
             defaultMode: 0600
         - name: {{ .Values.secretVolumeName }}
           secret:
-            secretName: "{{ coalesce .Values.msTeams.appSecretFromSecret (printf "%s-secret" (include "ms-teams.fullname" .)) }}"
+            secretName: "{{ coalesce .Values.msTeams.appSecretFromSecret (printf "%s-secret" (include "msteams.fullname" .)) }}"
             defaultMode: 0600
