+++ charts/access/msteams/templates/secret.yaml
@@ -3,7 +3,7 @@ apiVersion: v1
 kind: Secret
 type: Opaque
 metadata:
-  name: {{ include "ms-teams.fullname" . }}-secret
+  name: {{ include "msteams.fullname" . }}-secret
 data:
   {{ .Values.msTeams.appSecretFromSecretPath | quote }}: {{ .Values.msTeams.appSecret | b64enc }}
 {{- end }}
