+++ charts/access/msteams/tests/configmap_test.yaml
@@ -18,7 +18,7 @@ tests:
         "*":
           - dev-access-requests
       log:
-        output: /var/log/teleport-ms-teams.log
+        output: /var/log/teleport-msteams.log
         severity: DEBUG
     asserts:
       - matchSnapshot: {}
