+++ charts/access/msteams/tests/secret_test.yaml
@@ -12,7 +12,7 @@ tests:
   - it: should not exist when using external secret
     set:
       msTeams:
-        appSecretFromSecret: my-ms-teams-secret
+        appSecretFromSecret: my-msteams-secret
     asserts:
       - hasDocuments:
           count: 0
